Separate the represented persistence responsibility from the Session root (#609)
Seventh delivery of the refactor push. 105 methods move into 5 bounded
submodules under session/persistence: save, load, load_authority, plans and
commit. Every file is under the 1,000-line review budget.

session/mod.rs drops from 45,001 to 42,725 production lines. Across the seven
deliveries the root falls from 75,841 to 42,725 - 44 per cent retired into
navigable owners.

The finalization ledger, identity-hydration guard and unknown-COMMIT handling
established by #585 and #586 move with their methods; their ordering is
unchanged. Load authority sequencing is given its own module so the
begin/complete pair and the guard that proves represented state is
authoritative stay together rather than being buried in a general load file.

Save and load methods belonging to spell, aura, item, creature, gameobject,
quest, pet, loot, movement and transfer already moved with those families
under #589, #597, #599, #601, #603, #605 and #607, and are not touched again.

Safety established by comparison: the impl surface is 1,342 methods before and
after, an identical multiset, and the ownership baseline reports 3,761
associated items unchanged. All 18 baseline differences are private to
pub(in crate::session); nothing became public. The twelve direct-registry
deltas cover the identical method set, re-attributed to their new files.

Verified: cargo test -p wow-world --lib 3,822 passed and 0 failed,
session-ownership-check --syntax-only PASS, check_architecture.py check and
self-test PASS, fmt and diff hygiene clean. Ceiling tightened.

Real durability, crash and unknown-COMMIT recovery evidence is not produced
here and remains owned by #584.
